chore: rolls back simplificaiton to root layout
