+++ packages/next/src/layouts/Root/index.tsx
@@ -1,13 +1,20 @@
 import type { SanitizedConfig } from 'payload/types'
 
+import { auth } from '@payloadcms/next/utilities/auth'
+import { translations } from '@payloadcms/translations/client'
+import { RootProvider, buildComponentMap } from '@payloadcms/ui'
 import '@payloadcms/ui/scss/app.scss'
 import { headers as getHeaders } from 'next/headers'
 import { parseCookies } from 'payload/auth'
 import { createClientConfig } from 'payload/config'
+import { deepMerge } from 'payload/utilities'
 import React from 'react'
 import 'react-toastify/dist/ReactToastify.css'
 
 import { getRequestLanguage } from '../../utilities/getRequestLanguage'
+import { DefaultEditView } from '../../views/Edit/Default'
+import { DefaultListView } from '../../views/List/Default'
+import { DefaultCell } from '../../views/List/Default/Cell'
 
 export const metadata = {
   description: 'Generated by Next.js',
@@ -37,10 +44,33 @@ export const RootLayout = async ({
 
   const dir = rtlLanguages.includes(lang) ? 'RTL' : 'LTR'
 
+  const mergedTranslations = deepMerge(translations, clientConfig.i18n.translations)
+
+  const languageOptions = Object.entries(translations || {}).map(([language, translations]) => ({
+    label: translations.general.thisLanguage,
+    value: language,
+  }))
+
+  const componentMap = buildComponentMap({
+    DefaultCell,
+    DefaultEditView,
+    DefaultListView,
+    config,
+  })
+
   return (
     <html dir={dir} lang={lang}>
       <body>
-        {children}
+        <RootProvider
+          componentMap={componentMap}
+          config={clientConfig}
+          fallbackLang={clientConfig.i18n.fallbackLanguage}
+          lang={lang}
+          languageOptions={languageOptions}
+          translations={mergedTranslations[lang]}
+        >
+          {children}
+        </RootProvider>
         <div id="portal" />
       </body>
     </html>
